requirements.txt
feedparser==6.0.10
openai==1.3.0
python-telegram-bot==20.7
schedule==1.2.0
pandas==2.1.4
python-dotenv==1.0.0